+++ packages/react/src/components/sideNavigation/SideNavigation.stories.tsx
@@ -133,11 +133,7 @@ export const WithIcons = (args: SideNavigationProps) => {
       <div className="example-page">
         <div className="example-page-side-navigation">
           <SideNavigation {...args}>
-            <SideNavigation.MainLevel
-              id="main-level-link-1"
-              label="Main level accordion"
-              icon={<IconHome aria-hidden />}
-            >
+            <SideNavigation.MainLevel id="main-level-link-1" label="Main level accordion" icon={<IconHome />}>
               <SideNavigation.SubLevel
                 active={active === '/sub-level-1'}
                 id="sub-level-link-1"
@@ -153,11 +149,7 @@ export const WithIcons = (args: SideNavigationProps) => {
                 onClick={handleClick(setActive)}
               />
             </SideNavigation.MainLevel>
-            <SideNavigation.MainLevel
-              id="main-level-link-2"
-              label="Main level accordion"
-              icon={<IconHome aria-hidden />}
-            >
+            <SideNavigation.MainLevel id="main-level-link-2" label="Main level accordion" icon={<IconHome />}>
               <SideNavigation.SubLevel
                 active={active === '/sub-level-3'}
                 id="sub-level-link-3"
@@ -179,7 +171,7 @@ export const WithIcons = (args: SideNavigationProps) => {
               href="/main-level-3"
               label="Main level link"
               onClick={handleClick(setActive)}
-              icon={<IconHome aria-hidden />}
+              icon={<IconHome />}
             />
             <SideNavigation.MainLevel
               id="main-level-link-4"