+++ packages/react/src/icons/IconBox.tsx
@@ -4,24 +4,24 @@ import styles from './Icon.module.css';
 import { IconProps, IconSize } from './Icon.interface';
 
 export const IconBox = ({
-  ariaLabel = 'box',
-  ariaLabelledby,
-  ariaHidden = true,
+  'aria-label': ariaLabel = 'box',
+  'aria-hidden': ariaHidden = true,
   className = '',
   color,
   size = IconSize.Small,
   style = {},
-}: React.SVGProps<SVGSVGElement> & IconProps) => (
+  ...rest
+}: IconProps) => (
   <svg
+    aria-label={ariaLabel}
+    aria-hidden={ariaHidden}
     className={[styles.icon, styles[size], className].filter((e) => e).join(' ')}
     role="img"
     xmlns="http://www.w3.org/2000/svg"
     viewBox="0 0 24 24"
-    aria-label={ariaLabel}
-    aria-labelledby={ariaLabelledby}
-    aria-hidden={ariaHidden}
     color={color}
     style={style}
+    {...rest}
   >
     <path fillRule="evenodd" clipRule="evenodd" d="M20 7V21H4V7H20ZM18 9V19H6V9H18Z" fill="currentColor" />
     <path fillRule="evenodd" clipRule="evenodd" d="M2 3H22V9H2V3ZM4 5V7H20V5H4Z" fill="currentColor" />